This allows the checking of username and password
This allows the checking of username and password. Read through the code to figure out how it works. there are 2 comparision methods that compare the username and password. One is slightly more advanced but it will run faster the other is with the normal if satements. I commented out the one with if statements but you can chosse whichever one you want.